// cypress/support/typings/userdetails.d.ts
interface UserDetails {
    user: string;
    password: string;
    email: string;
    title: string;
    firstName: string;
    lastName: string;
    dob: string;
    company: string;
    address: string;
    city: string;
    state: string;
    country: string;
    zipcode: string;
    mobile: string;
}

export default UserDetails